requirements.txt
torch
torchvision
transformers
openai
dashscope
pillow
pandas
numpy
open_clip_torch